// src/pages/inhouse-disposition/parts.type.ts
export type PartId =
  | "P1"
  | "P2"
  | "P3"
  | "E4"
  | "E5"
  | "E6"
  | "E7"
  | "E8"
  | "E9"
  | "E10"
  | "E11"
  | "E12"
  | "E13"
  | "E14"
  | "E15"
  | "E16"
  | "E17"
  | "E18"
  | "E19"
  | "E20"
  | "E26"
  | "E29"
  | "E30"
  | "E31"
  | "E49"
  | "E50"
  | "E51"
  | "E54"
  | "E55"
  | "E56";